Move the whole atlas generation code into Common/Render
