+++ Common/Render/AtlasGen.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int GenerateFromScript(const char *script_file, const char *atlas_name, bool highcolor);
