+++ Common/Render/kanjifilter.h
